CHANGE: refactored guaranteed_setting_service.rb to simplify and fix it.  added mocks and refactored specs
