+++ app/controllers/concerns/bravura_template_base/base_controller_concern.rb
@@ -1,4 +1,5 @@
 # app/controllers/concerns/bravura_template_base/base_controller_concern.rb
+
 require "pagy"
 
 module BravuraTemplateBase
